+++ test/cram/lint/migrate.t
@@ -15,14 +15,15 @@ and the error message points to the corresponding objective.
   > "#1058","Application and Operational Metrics","Complete ✅","Q4 2023 - Oct - Dec","Ci & Ops","QA","Operational Metrics for Core OCaml Services","Jane Street - Community","pillar/qa","50."
   > "#1090","Property-Based Testing for Multicore","Active 🏗","Q1 2024 - Jan - Mar","Compiler and language","Compiler","Property-Based Testing for Multicore","","pillar/compiler,team/compiler&language,Proposal","25."
   > "#1115","General okra maintenance","Draft","","","","Maintenance - internal tooling","","pillar/ecosystem,team/internal-tooling",""
+  > "#2000","JSOO Effect Performance","","","","","JSOO Effect Performance","","",""
   > EOF
 
   $ cat > admin/data/team-objectives.csv << EOF
-  > "id","title","status","quarter","team","pillar","objective","funder","labels","progress"
-  > "#548","CAstore: persistent and distributed heap for OCaml","In Progress","Q2 2024","Irmin","Ecosystem","","Jane Street","Proposal",""
-  > "#558","Property-Based Testing for Multicore","In Progress","Q2 2024","Compiler & Language","Compiler","","","Proposal",""
-  > "#677","Improve OCaml experience on Windows","Todo","Q2 2024","Multicore applications","Ecosystem","","","",""
-  > "#701","JSOO Effect Performance","","Q2 2024","Compiler & Language","Compiler","","","focus/technology,level/team",""
+  > "id","title","status","start on quarter","end on quarter","team","pillar","objective","funder","labels","progress"
+  > "#548","CAstore: persistent and distributed heap for OCaml","In Progress","Q2 2024","Q2 2024","Irmin","Ecosystem","","Jane Street","Proposal",""
+  > "#558","Property-Based Testing for Multicore","In Progress","Q2 2024","Q2 2024","Compiler & Language","Compiler","","","Proposal",""
+  > "#677","Improve OCaml experience on Windows","Todo","Q2 2024","Q2 2024","Multicore applications","Ecosystem","","","",""
+  > "#701","JSOO Effect Performance","","Q4 2024","Q4 2024","Compiler & Language","Compiler","","","focus/technology,level/team",""
   > EOF
 
   $ cat > weekly.md << EOF
@@ -129,3 +130,20 @@ The whole title is taken into account (there was an issue with `:`)
 
   $ okra lint -e -C admin admin/weekly/2024/24/eng1.md
   [OK]: admin/weekly/2024/24/eng1.md
+
+Using a work-item that has a parent objective, but which start/end quarter are invalid.
+The parent objective is not suggested in the error message:
+
+  $ cat > admin/weekly/2024/24/eng1.md << EOF
+  > # Last week
+  > 
+  > - JSOO Effect Performance (#2000)
+  >   - @eng1 (5 days)
+  >   - Something
+  > EOF
+
+  $ okra lint -e -C admin admin/weekly/2024/24/eng1.md
+  File "admin/weekly/2024/24/eng1.md", line 3:
+  Error: Invalid objective:
+         "JSOO Effect Performance (#2000)" is a work-item. You should use an objective instead.
+  [1]